EDM-569: Rename inputs to options
