+++ tests/unit/plugins/module_utils/test_api_module.py
@@ -10,7 +10,7 @@ from tests.unit.utils import set_module_args
 from plugins.module_utils.api_module import FlightctlAPIModule
 from plugins.module_utils.constants import Kind
 from plugins.module_utils.exceptions import FlightctlException
-from plugins.module_utils.inputs import ApprovalInput
+from plugins.module_utils.options import ApprovalOptions
 
 
 @pytest.fixture
@@ -43,7 +43,7 @@ def test_approve_success(api_module):
     mock_request = Mock(return_value=mock_response)
     api_module.request = mock_request
 
-    input = ApprovalInput(Kind.ENROLLMENT, "test-enrollment", True)
+    input = ApprovalOptions(Kind.ENROLLMENT, "test-enrollment", True)
     api_module.approve(input)
     mock_request.assert_called_with("POST", "https://test-flightctl-url.com/api/v1/enrollmentrequests/test-enrollment/approval", **input.to_request_params())
 
@@ -55,7 +55,7 @@ def test_approve_404(api_module):
     mock_request = Mock(return_value=mock_response)
     api_module.request = mock_request
 
-    input = ApprovalInput(Kind.ENROLLMENT, "test-enrollment", True)
+    input = ApprovalOptions(Kind.ENROLLMENT, "test-enrollment", True)
     with pytest.raises(FlightctlException, match="Unable to approve EnrollmentRequest for test-enrollment"):
         api_module.approve(input)
 
@@ -66,7 +66,7 @@ def test_approve_csr(api_module):
     mock_request = Mock(return_value=mock_response)
     api_module.request = mock_request
 
-    input = ApprovalInput(Kind.CSR, "test-csr", True)
+    input = ApprovalOptions(Kind.CSR, "test-csr", True)
     api_module.approve(input)
     mock_request.assert_called_with("POST", "https://test-flightctl-url.com/api/v1/certificatesigningrequests/test-csr/approval", **input.to_request_params())
 
@@ -77,6 +77,6 @@ def test_deny_csr(api_module):
     mock_request = Mock(return_value=mock_response)
     api_module.request = mock_request
 
-    input = ApprovalInput(Kind.CSR, "test-csr", False)
+    input = ApprovalOptions(Kind.CSR, "test-csr", False)
     api_module.approve(input)
     mock_request.assert_called_with("DELETE", "https://test-flightctl-url.com/api/v1/certificatesigningrequests/test-csr/approval")
